% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalms.R
\docType{data}
\name{datalms5_19}
\alias{datalms5_19}
\title{This is a look up table data for internal use only.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 482 rows and 5 columns.
}
\usage{
datalms5_19
}
\description{
This is a look up table data for internal use only.
}
\keyword{data}
\keyword{internal}
\keyword{look}
\keyword{up}
